+++ tsconfig.json
@@ -6,5 +6,5 @@
     "declaration": true,
     "outDir": "./dist"
   },
-  "exclude": ["tests", "dist", "node_modules", "altceva.ts"]
+  "exclude": ["tests", "dist", "node_modules"]
 }
